Refactor into hardware and software checks, improve make cmds

+++ miner/src/checks/hardware/storage.rs
@@ -21,9 +21,9 @@ pub fn get_storage_info() -> Result<(u64, u64), SystemCheckError> {
             }
         }?;
 
-        let total = statvfs.f_blocks * statvfs.f_frsize;
-        let free = statvfs.f_bavail * statvfs.f_frsize;
-        Ok((total as u64, free as u64))
+        let total = (statvfs.f_blocks as u64) * (statvfs.f_frsize as u64);
+        let free = (statvfs.f_bavail as u64) * (statvfs.f_frsize as u64);
+        Ok((total, free))
     }
 
     #[cfg(not(unix))]
